Make TV API service configurable.
It's possible to configure the type of implementation used for the
TV API. Available options include mock (default), vlcdvb, coolstream and
coolstream_rest. The vlcdvb impl also has a configurable path setting
for the channels.conf, default channels.conf is for Berlin.
 E.g. to use the vlcdvb impl with a custom channels.conf use:
{name: "tv", params: {impl: "vlcdvb", path: "/path/to/chan.conf"}}
when loading the module in the PZP.

Jira issue: WP-32