Add RPC registry related functions to a seperate file.
Prepare to move the discovery related code to its own file by factoring
out the registry related code to its own file/module. This registry code
is to be used by RPC and discovery later on, both will have a reference
to the registry object. Also instantiate the in PZH and PZP for later
use.

Jira-issue: WP-84