Remove fromObjectRef parameter from JSON-RPC messages.
FromObjectRef was used to associate the RPC object and a callback object
on the request issueing side. Now there is no separate callback object and
instead, the RPC object itself is used to register methods on it that can
receive requests from the responding side. The id of the RPC object is
used to register it as a service. The request receiving side still gets
an objectRef parameter which is now an object looking like this:

objectRef = {
    rpcId: "some id",
    from: "address the request came from"
}

That object can be used with createRPC as before to call methods on the
caller object of the requesting side. Since it now also includes the from
address, the RPC handler doesn't have to save that internally, so the map
for that (objRefCacheTable) is removed. That map was getting bigger since
nothing could be removed from it. The JSON-RPC messages now have no
paramters outside the 2.0 specification.

Jira-issue: WP-180