Return negative result when CertificateManager fails.
The Certificate manager code was returning '0' despite reporting errors.  This
change returns an error code which is caught in JavaScript.  The tests have
been updated to try using an invalid subjectAltName and to show that the
error propagates appropriately.  Also, tests have been fixed for newer OpenSSL
versions which check the subjectAltName format.

Jira-Issue: WP-403