Reorder creation of platform socket object and script injection
Jira issue: WP-456

The addition of the WebinosSocket object is moved to be prior to
page loading, and the injection of the associated script is moved
to the start of page loading.

In this revision, the injection of the webinos.js script is
temporarily disabled, since the existing mechanism is limited
to strings of length 65535 chars (and webinos.js is now too long).