Do not register callback rpc objects in registry
Instead register them inside the rpcHandler. RPC supports two
different type of requests. 1) Normal requests to services registered in
the registry. 2) Requests to callback objects that function as replies to
an initial request from that callback object. These callback objects are
registered with registerCallbackObject method.

This change makes sure the callback objects don't pollute the registry.
Services from the registry need to be communicated to other entities in
the PZ. While callback objects are completely private to an entity, so
information about them should never be sent anywhere.

Jira-issue: WP-513