Use webinos-rpc instead of in tree rpc
rpc code was factored out into its own node module. code for requiring
the api service modules (from in tree) was factored out into the utility
code and is not longer part of the rpc code. the separate webinos-rpc
module cannot do this as the api services are in different directory trees
now.

Jira-issue: WP-258