New installer for Android
Jira issue: WP-597

Add new installer app for Android that contains the wrt and app application
packages as assets. The ant build file has been updated to build all three
apps, and package the built apks into the installer.

Platform initialisation (extracting the packaged modules on startup) is also
significantly revised in this commit, to trigger initialisation on install,
and also ensure that a progress dialog is shown in the runtime if platform
init is still pending on launch.