Fixed sensor/iot register service defects + fix of command signatures
Jira-issue: WP-712

There was some defects in iot driver design which meant that if
a driver tried to register directly in the init function this
would fail. The following had to be done:

- Loading and initializing drivers is done in 2 separate steps since
a driver can contain both sensor and/or actuators and both listeners
has to be initialized.

- Driverinterface is created and then as a separate function call
the module is connected. otherwise the sensor/actuator module haven't
stored the driverinterface reference when registrations are called
back.

Also, some changes has been made to some of the driverinterface
command signatures to better support the needed use-cases:

- An error callback has been added to driverinterface commands
so that drivers can report errors back.

- The hard-coded eventmode parameter sent in 'start' command data
has been removed. In accordance with the Sensor API this is sent
in configuration data instead.