Additional policy tests in jasmine directories, with the working tests added as an additional unit test.
This commit consists of an update to the unit test script. This update invokes the directory of working test.
In addition there are two new directories of jasmine tests.
The first directory is
webinos/core/manager/policy_manager/test/jasmine.policy.tests.working
which  contains the working tests that have been added to unit-test.sh. These tests exercise various aspects of the policy manager.

The second directory is
webinos/core/manager/policy_manager/test/jasmine.policy.tests.broken
contains a collection of jasmine tests that highlight issues with how policies are evaluated.

This commit is a set of tests only and DOES NOT fix jira issus WP-912.
The primary error highlighted by the tests in the broken directory is that conditions containing subject matches do not behave as expected.

The tests can be invoked manually by using cd to navigate the the directory and executing "jasmine-node ." from inside the directory.

jira issue: WP-912